scripts/requirements.txt
requests
pandas
python-dotenv
gitpython